[update statistic unit number to float]

+++ src/main/java/com/example/heroku/model/statistics/BenifitByProduct.java
@@ -13,6 +13,6 @@ public class BenifitByProduct {
     private String product_unit_second_id;
     private float revenue;
     private float profit;
-    private int number_unit;
+    private float number_unit;
     protected Timestamp createat;
 }
